+++ DbCreator/Sql/3.sql
@@ -1,5 +1,5 @@
 ﻿Create table Attendee(
-Id int identity(1,1) not null,
+Id bigint identity(1,1) not null,
 FullName nvarchar(255) not null,
 PhoneNumber nvarchar(255),
 Email nvarchar(255),
@@ -9,7 +9,7 @@ primary key (Id)
 );
 
 Create table AcademicProgram(
-Id int identity(1,1) not null,
+Id bigint identity(1,1) not null,
 Name nvarchar(255) not null,
 primary key (Id)
 );	
@@ -17,9 +17,9 @@ primary key (Id)
 Create table Pupil(
 Sex int not null,
 YearOfGraduation int not null,
-EnterProgram_Id int,
-School_Id int not null,
-Attendee_Id int not null,
+EnterProgram_Id bigint,
+School_Id bigint not null,
+Attendee_Id bigint not null,
 
 constraint FK_Pupil_EnterProgram foreign key (EnterProgram_Id) references AcademicProgram(Id),
 constraint FK_Pupil_School foreign key (School_Id) references School(Id),
@@ -28,8 +28,8 @@ constraint AK_AttendeeId unique(Attendee_Id)
 );
 
 Create table IntrestingProgram(
-AcademicProgram_Id int not null,
-Pupil_Id int not null,
+AcademicProgram_Id bigint not null,
+Pupil_Id bigint not null,
 
 constraint FK_IntrestingProgram_AcademicProgram foreign key (AcademicProgram_Id) references AcademicProgram(Id),
 constraint FK_IntrestingProgram_Pupil foreign key (Pupil_Id) references Pupil(Attendee_Id)
@@ -37,8 +37,8 @@ constraint FK_IntrestingProgram_Pupil foreign key (Pupil_Id) references Pupil(At
 );	
 
 Create table RegistrarionProgram(
-AcademicProgram_Id int not null,
-Pupil_Id int not null,
+AcademicProgram_Id bigint not null,
+Pupil_Id bigint not null,
 
 constraint FK_RegistrarionProgram_AcademicProgram foreign key (AcademicProgram_Id) references AcademicProgram(Id),
 constraint FK_RegistrarionProgram_Pupil foreign key (Pupil_Id) references Pupil(Attendee_Id)
